moving the interface and updating it
Moved FirebaseInterface into the src folder. Added docblocks. Added the
firebase namespace.
